# AOT Module Simplified - Low Overhead version of AOT Module